Order of service start log statements
In the case of services which are overridden to always use release
versions, the console out logging can be slightly misleading as to which
version actually started up.

ie.
Service 'MYSERVICE' has been overridden to always use 'RELEASE' version
Starting 'MYSERVICE' from SNAPSHOT...

This logging is trying to indicate that a snapshot was request but a
release version was actually started. Reversing the order which these
lines appear in the console log would make it clearer that that actually
happened.

ie.
Starting 'MYSERVICE' from SNAPSHOT...
Service 'MYSERVICE' has been overridden to always use 'RELEASE' version