Allow later patch versions of argcomplete
Using ~ should install 1.12.1 if available (which it is)

Specifically, we want the changes here:
https://github.com/kislyuk/argcomplete/issues/321

importlib-metadata 2.0 was released, and argcomplete defined that 2.0
should never be use that version, until the most recent current version.

I believe that one of our test dependencies imported a higher version of
argcomplete before the resolver realised this. There is potentially
another piece of work to NOT install test dependencies for users of
servicemanager, but that can be postponed for now.